Move Charset to package encoding.charset
